Add DraftProjectDTO
